requirements.txt
torch
torchvision
flask
flask-cors
pillow
numpy
matplotlib
scikit-learn
tqdm
opencv-python
albumentations